Fix: frequency is calculated as a double
So it is better to take it as a double than implicit-cast to uint64_t
and then back to double. That cast roundtrip makes no sense.

+++ new/src/tools/exec/vm.cpp
@@ -88,7 +88,7 @@ auto format_time(std::chrono::microseconds const us) -> std::string
     }
     return out.str();
 }
-auto format_hz(uint64_t const hz) -> std::string
+auto format_hz(double const hz) -> std::string
 {
     auto out = std::ostringstream{};
     if (hz > 1e3) {
